requirements.txt
flask
flask_cors
torch
torchaudio
torchvision
pandas
nltk>=3.6.1
plotly
joblib
wordcloud
matplotlib
mysql-connector-python
selenium
scikit-learn
amazoncaptcha
googletrans
webdriver_manager
transformers
scipy
google.generativeai